% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-before.R
\name{before_last_dot}
\alias{before_last_dot}
\alias{str_before_last_dot}
\title{Get the part of a string before the last period.}
\usage{
before_last_dot(...)

str_before_last_dot(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_before_last_dot]{strex::str_before_last_dot()}}.
}
