% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{BeforeLastDot}
\alias{BeforeLastDot}
\title{Get the part of a string before the last period.}
\usage{
BeforeLastDot(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector.
}
\description{
This is usually used to get the part of a file name that doesn't include the
file extension. It is vectorised over \code{string}.
}
\examples{
BeforeLastDot(c("spreadsheet1.csv", "doc2.doc"))

}
