% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{can_be_numeric}
\alias{can_be_numeric}
\title{Check if a string could be considered as numeric.}
\usage{
can_be_numeric(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector. \code{TRUE} if the argument can be considered to be
numeric or \code{FALSE} otherwise.
}
\description{
After padding is removed, could the input string be considered to be numeric,
i.e. could it be coerced to numeric. This function is vectorized over its one
argument.
}
\examples{
can_be_numeric("3")
can_be_numeric("5 ")
can_be_numeric(c("1a", "abc"))
}
