% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{CanBeNumeric}
\alias{CanBeNumeric}
\title{Check if a string could be considered as numeric.}
\usage{
CanBeNumeric(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector. \code{TRUE} if the argument can be considered to be numeric or \code{FALSE}
otherwise.
}
\description{
After padding is removed, could the input string be considered to be numeric,
i.e. could it be coerced to numeric. This function is vectorised over its one
argument.
}
\examples{
CanBeNumeric("3")
CanBeNumeric("5 ")
CanBeNumeric(c("1a", "abc"))
}
