% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CharListElemsNthElem}
\alias{CharListElemsNthElem}
\alias{NumListElemsNthElem}
\title{Get the nth element of each vector in a list.}
\usage{
CharListElemsNthElem(char_list, n)

NumListElemsNthElem(num_list, n)
}
\arguments{
\item{char_list}{A list of character vectors.}

\item{n}{The index of the element that you want from each vector.}

\item{num_list}{A list of numeric vectors.}
}
\value{
A list.
}
\description{
Get the nth element of each vector in a list.
}
\examples{
CharListElemsNthElem(list(c("a", "b", "c"), c("d", "f", "a")), 2)
NumListElemsNthElem(list(1:5, 0:2), 4)
}
