% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{create_dir}
\alias{create_dir}
\title{Create directories if they don't already exist}
\usage{
create_dir(...)
}
\arguments{
\item{...}{The names of the directories, specified via relative or absolute
paths. Duplicates are ignored.}
}
\value{
Invisibly, a vector with a \code{TRUE} for each time a directory was
actually created and a \code{FALSE} otherwise. This vector is named with the
paths of the directories that were passed to the function.
}
\description{
Given the names of (potential) directories, create the ones that do not
already exist.
}
\examples{
\dontrun{
create_dir(c("mydir", "yourdir"))
remove_dir(c("mydir", "yourdir"))
}
}
