% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{create_dirs}
\alias{create_dirs}
\title{Create directories if they don't already exist}
\usage{
create_dirs(...)
}
\arguments{
\item{...}{The names of the directories, specified via relative or absolute
paths. Duplicates are ignored.}
}
\value{
Invisibly, a vector with a \code{TRUE} for each time a directory was
actually created and a \code{FALSE} otherwise. Theis vector is named with the
paths of the directories that were passed to the function.
}
\description{
Given the names of (potential) directories, create the ones that do not
already exist.
}
\examples{
setwd(tempdir())
create_dirs(c("mydir", "yourdir"))
remove_dirs(c("mydir", "yourdir"))
}
