% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-currency.R
\name{currency}
\alias{currency}
\alias{str_extract_currencies}
\alias{extract_currencies}
\alias{str_nth_currency}
\alias{nth_currency}
\alias{str_first_currency}
\alias{first_currency}
\alias{str_last_currency}
\alias{last_currency}
\title{Get the currencies of numbers within a string.}
\usage{
str_extract_currencies(...)

extract_currencies(...)

str_nth_currency(...)

nth_currency(...)

str_first_currency(...)

first_currency(...)

str_last_currency(...)

last_currency(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
See \code{\link[strex:currency]{strex::str_extract_currencies()}}.
}
