% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-pad.R
\name{extend_char_vec}
\alias{extend_char_vec}
\title{Pad a character vector with empty strings.}
\usage{
extend_char_vec(char_vec, extend_by = NA, length_out = NA)
}
\arguments{
\item{char_vec}{A character vector. The thing you wish to expand.}

\item{extend_by}{A non-negative integer. By how much do you wish to extend
the vector?}

\item{length_out}{A positive integer. How long do you want the output vector
to be?}
}
\value{
A character vector.
}
\description{
Extend a character vector by appending empty strings at the end.
}
\examples{
extend_char_vec(1:5, extend_by = 2)
extend_char_vec(c("a", "b"), length_out = 10)
}
