% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-extract-non-nums.R
\name{extract_non_numerics}
\alias{extract_non_numerics}
\alias{str_extract_non_numerics}
\alias{nth_non_numeric}
\alias{str_nth_non_numeric}
\alias{first_non_numeric}
\alias{str_first_non_numeric}
\alias{last_non_numeric}
\alias{str_last_non_numeric}
\title{Extract non-numbers from a string.}
\usage{
extract_non_numerics(...)

str_extract_non_numerics(...)

nth_non_numeric(...)

str_nth_non_numeric(...)

first_non_numeric(...)

str_first_non_numeric(...)

last_non_numeric(...)

str_last_non_numeric(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copies of \code{\link[strex:str_extract_non_numerics]{strex::str_extract_non_numerics()}} and friends.
}
