% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-extract-nums.R
\name{extract_numbers}
\alias{extract_numbers}
\alias{str_extract_numbers}
\alias{nth_number}
\alias{str_nth_number}
\alias{first_number}
\alias{str_first_number}
\alias{last_number}
\alias{str_last_number}
\title{Extract numbers from a string.}
\usage{
extract_numbers(...)

str_extract_numbers(...)

nth_number(...)

str_nth_number(...)

first_number(...)

str_first_number(...)

last_number(...)

str_last_number(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copies of \code{\link[strex:str_extract_numbers]{strex::str_extract_numbers()}} and friends.
}
