% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesstrings.R
\name{filesstrings-defunct}
\alias{filesstrings-defunct}
\title{Defunct functions}
\arguments{
\item{...}{Defunct function arguments.}
}
\description{
These functions have been made defunct, mostly because the naming style of
the package has been changed. Some have been removed because they were not
well-done.
}
