% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesstrings.R
\docType{package}
\name{filesstrings}
\alias{filesstrings}
\alias{filesstrings-package}
\title{\code{filesstrings}: handy file and string manipulation}
\description{
Convenient functions for moving files, deleting directories, and a variety of
string operations that facilitate manipulating files and extracting
information from strings.
}
\references{
Rory Nolan and Sergi Padilla-Parra (2017). filesstrings: An R
package for file and string manipulation. The Journal of Open Source
Software, 2(14).  \doi{10.21105/joss.00260}.
}
