% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{FirstNumber}
\alias{FirstNumber}
\alias{LastNumber}
\title{The first number within a string}
\usage{
FirstNumber(string, leave.as.string = FALSE, err.na = TRUE)

LastNumber(string, leave.as.string = FALSE, err.na = FALSE)
}
\arguments{
\item{string}{The String from which to extract the first number.}

\item{leave.as.string}{Do you want to return the number as a string
(\code{TRUE}) or as numeric (\code{FALSE}, the default)?}

\item{err.na}{If no first number is found (because there isn't one), should
the function return \code{NA} (\code{TRUE}, the default), or should it spit
out an error?}
}
\value{
The first number in \code{string}, or \code{NA} if no number is found
  and \code{err.na} is \code{TRUE}.
}
\description{
Extract the first (or last) contiguous sequence of \emph{numeric characters}
from a string (where only the digits 0-9 are considered to be numeric
characters and hence "." and "-" are not). Please view the examples at the
bottom of this page to ensure that you understand how this function works,
and its limitations.
}
\examples{
FirstNumber("abc123abc456")  # returns 123
FirstNumber("abc1.23abc456")  # returns 1, not 1.23
FirstNumber("-123abc456")  # returns 123, not -123
}

