% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{GetCurrencies}
\alias{GetCurrencies}
\title{Get the currencies of numbers within a string.}
\usage{
GetCurrencies(string)
}
\arguments{
\item{string}{}
}
\value{
A data frame with one column for the currency symbol and one for the
  amount.
}
\description{
The currency is defined as the character coming before the number in the
string. If nothing comes before (i.e. if the number is the first thing in the
string), the currency can be the empty string, similarly the currency can be
a space, comma or any manner of thing. This function does not allow for
leading decimal points.
}
\examples{
GetCurrencies("35.00 $1.14 abc5 £3.8 77")
}

