% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mat2RowList}
\alias{Mat2ColList}
\alias{Mat2RowList}
\title{List of matrix rows or columns.}
\usage{
Mat2RowList(mat)

Mat2ColList(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\description{
From a matrix, create a list whose \eqn{i}th element is the \eqn{i}th row or
column of the matrix.
}
\examples{
m <- matrix(1:6, nrow = 2)
print(m)
Mat2RowList(m)
Mat2ColList(m)
}

