% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg-match.R
\name{match_arg}
\alias{match_arg}
\alias{str_match_arg}
\title{Argument Matching}
\usage{
match_arg(...)

str_match_arg(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_match_arg]{strex::match_arg()}}.
}
