% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-num-before.R
\name{nth_number_before_mth}
\alias{nth_number_before_mth}
\alias{str_nth_number_before_mth}
\alias{nth_number_before_first}
\alias{nth_number_before_last}
\alias{first_number_before_mth}
\alias{last_number_before_mth}
\alias{first_number_before_first}
\alias{first_number_before_last}
\alias{last_number_before_first}
\alias{last_number_before_last}
\alias{str_nth_number_before_first}
\alias{str_nth_number_before_last}
\alias{str_first_number_before_mth}
\alias{str_last_number_before_mth}
\alias{str_first_number_before_first}
\alias{str_first_number_before_last}
\alias{str_last_number_before_first}
\alias{str_last_number_before_last}
\title{Find the \code{n}th number before the \code{m}th occurrence of a pattern.}
\usage{
nth_number_before_mth(...)

str_nth_number_before_mth(...)

nth_number_before_first(...)

nth_number_before_last(...)

first_number_before_mth(...)

last_number_before_mth(...)

first_number_before_first(...)

first_number_before_last(...)

last_number_before_first(...)

last_number_before_last(...)

str_nth_number_before_first(...)

str_nth_number_before_last(...)

str_first_number_before_mth(...)

str_last_number_before_mth(...)

str_first_number_before_first(...)

str_first_number_before_last(...)

str_last_number_before_first(...)

str_last_number_before_last(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_nth_number_before_mth]{strex::str_nth_number_before_mth()}}.
}
