% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{NthNumber}
\alias{NthNumber}
\title{Extract the \eqn{n}th number from a string.}
\usage{
NthNumber(string, n, err.na = FALSE, leave.as.string = FALSE)
}
\arguments{
\item{string}{The string from which to extract the \eqn{n}th number.}

\item{n}{A natural number, the index of the number we wish to extract from
\code{string}}

\item{err.na}{If there is an error (most likely due to there not being a
\eqn{n}th number), if you would like to return NA instead of giving an
error (the default), set this to \code{TRUE}.}

\item{leave.as.string}{Do you want to return the number as a string
(\code{TRUE}) or as numeric (\code{FALSE}, the default)?}
}
\value{
A number.
}
\description{
Extract the \eqn{n}th number from a string, where a \emph{number} is defined
as a contiguous sequence of \emph{numeric characters} from a string (where
only the digits 0-9 are considered to be numeric characters and hence "."
and "-" are not). Please view the examples at the bottom of this page to
ensure that you understand how this function works, and its limitations.
}
\examples{
NthNumber("abc123def456ghi789jkl", 2)
NthNumber("abc1.23abc456", 1)
NthNumber("-123abc456", 1)
}

