% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PasteCollapseListElems}
\alias{PasteCollapseListElems}
\title{Apply paste collapse to each element of a list.}
\usage{
PasteCollapseListElems(char_list, collapse = "")
}
\arguments{
\item{char_list}{A list of character vectors.}

\item{collapse}{See \code{\link{paste}}.}
}
\value{
A list of character vectors.
}
\description{
This is the same as doing
\code{sapply(char.list, paste, collapse = collapse)}, it's just faster.
}
\examples{
PasteCollapseListElems(list(1:3, c("a", 5, "rory")), collapse = "R")
}

