% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{PutFilesInDir}
\alias{PutFilesInDir}
\title{Put files in a directory}
\usage{
PutFilesInDir(file.names, dir.name)
}
\arguments{
\item{file.names}{character vector of file names, relative or absolute paths}

\item{dir.name}{The name of the directory into which to move the files,
relative or absolute path}
}
\value{
A logical vector with a \code{TRUE} for each successful file move and
a \code{FALSE} otherwise.
}
\description{
Move files to a directory, creating the directory if it does not already
exist.
}
