% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-remove.R
\name{remove_quoted}
\alias{remove_quoted}
\alias{str_remove_quoted}
\title{Remove the quoted parts of a string.}
\usage{
remove_quoted(string)

str_remove_quoted(string)
}
\arguments{
\item{string}{A character vector.}
}
\description{
See \code{\link[strex:str_remove_quoted]{strex::str_remove_quoted()}}.
}
