% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{RemoveFileSpaces}
\alias{RemoveFileSpaces}
\title{Remove spaces in file names}
\usage{
RemoveFileSpaces(dir = ".", pattern = "", replace.with = "")
}
\arguments{
\item{dir}{The directory in which to perform the operation.}

\item{pattern}{A regular expression. If specified, only files matching this
pattern wil be treated.}
}
\value{
A logical vector indicating which operation succeeded for each of the
  files attempted. Using a missing value for a file or path name will always
  be regarded as a failure.
}
\description{
Remove spaces in file names in a specified directory, replacing them with
whatever you want, default nothing.
}

