% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{StopOrNA}
\alias{StopOrNA}
\title{Stop or return NA}
\usage{
StopOrNA(stop, message = "")
}
\arguments{
\item{stop}{\code{TRUE} if you want to stop with an error message, \code{FALSE} if you want to return NA}

\item{message}{The error message to display if \code{stop = TRUE}}
}
\description{
Stop or return NA
}

