% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_elem}
\alias{str_elem}
\title{Extract a single character of a string, using its index.}
\usage{
str_elem(string, index)
}
\arguments{
\item{string}{A string.}

\item{index}{An integer. Negative indexing is allowed as in
\code{\link[stringr:str_sub]{stringr::str_sub()}}.}
}
\value{
A one-character string.
}
\description{
Extract a single character of a string, using its index.
}
\examples{
str_elem("abcd", 3)
str_elem("abcd", -2)
}
