% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-elem.R
\name{str_elems}
\alias{str_elems}
\alias{elems}
\title{Extract several single elements from a string.}
\usage{
str_elems(...)

elems(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_elems]{strex::str_elems()}}.
}
