% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-give-ext.R
\name{str_give_ext}
\alias{str_give_ext}
\alias{give_ext}
\title{Ensure a file name has the intended extension.}
\usage{
str_give_ext(...)

give_ext(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_give_ext]{strex::str_give_ext()}}.
}
