% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_list_nth_elems}
\alias{str_list_nth_elems}
\alias{num_list_nth_elems}
\title{Get the nth element of each vector in a list of numeric or character
vectors.}
\usage{
str_list_nth_elems(char_list, n)

num_list_nth_elems(num_list, n)
}
\arguments{
\item{char_list}{A list of character vectors.}

\item{n}{The index of the element that you want from each vector.}

\item{num_list}{A list of numeric vectors.}
}
\value{
A list.
}
\description{
These are faster implementations of procedures that could very easily be
done with \link[base:sapply]{base::sapply}.
}
\examples{
str_list_nth_elems(list(c("a", "b", "c"), c("d", "f", "a")), 2)
num_list_nth_elems(list(1:5, 0:2), 4)
}
