% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_list_remove_empties}
\alias{str_list_remove_empties}
\title{Remove empty strings from a character list.}
\usage{
str_list_remove_empties(char_list)
}
\arguments{
\item{char_list}{A list of character vectors.}
}
\value{
A list of character vectors.
}
\description{
Remove empty strings from a character list.
}
\examples{
str_list_remove_empties(list(c("a", "", "b"), "gg", c("", 1, "")))
}
