% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-locate.R
\name{str_locate_nth}
\alias{str_locate_nth}
\alias{locate_nth}
\alias{str_locate_first}
\alias{locate_first}
\alias{str_locate_last}
\alias{locate_last}
\title{Get the indices of the \eqn{n}th instance of a pattern.}
\usage{
str_locate_nth(...)

locate_nth(...)

str_locate_first(...)

locate_first(...)

str_locate_last(...)

locate_last(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_locate_nth]{strex::str_locate_nth()}}.
}
