% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-nice-nums.R
\name{str_nice_nums}
\alias{str_nice_nums}
\alias{nice_nums}
\alias{str_alphord_nums}
\alias{alphord_nums}
\title{Make string numbers comply with alphabetical order.}
\usage{
str_nice_nums(...)

nice_nums(...)

str_alphord_nums(...)

alphord_nums(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_alphord_nums]{strex::str_alphord_nums()}}.
}
