% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-elem.R
\name{str_paste_elems}
\alias{str_paste_elems}
\alias{paste_elems}
\title{Extract bits of a string and paste them together.}
\usage{
str_paste_elems(...)

paste_elems(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_paste_elems]{strex::str_paste_elems()}}.
}
