% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-remove.R
\name{str_remove_quoted}
\alias{str_remove_quoted}
\alias{remove_quoted}
\title{Remove the quoted parts of a string.}
\usage{
str_remove_quoted(...)

remove_quoted(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_remove_quoted]{strex::str_remove_quoted()}}.
}
