% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-singleize.R
\name{str_singleize}
\alias{str_singleize}
\alias{singleize}
\title{Remove back-to-back duplicates of a pattern in a string.}
\usage{
str_singleize(...)

singleize(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_singleize]{strex::str_singleize()}}.
}
