% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-split-by-nums.R
\name{str_split_by_nums}
\alias{str_split_by_nums}
\alias{split_by_nums}
\alias{split_by_numbers}
\alias{str_split_by_numbers}
\title{Split a string by its numeric characters.}
\usage{
str_split_by_nums(...)

split_by_nums(...)

split_by_numbers(...)

str_split_by_numbers(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_split_by_numbers]{strex::str_split_by_numbers()}}.
}
