% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-split.R
\name{str_split_camel_case}
\alias{str_split_camel_case}
\alias{split_camel_case}
\title{Split a string based on CamelCase}
\usage{
str_split_camel_case(string, lower = FALSE)

split_camel_case(string, lower = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{lower}{Do you want the output to be all lower case (or as is)?}
}
\description{
See \code{\link[strex:str_split_camel_case]{strex::str_split_camel_case()}}.
}
