% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-to-vec.R
\name{str_to_vec}
\alias{str_to_vec}
\alias{to_vec}
\title{Convert a string to a vector of characters}
\usage{
str_to_vec(...)

to_vec(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_to_vec]{strex::str_to_vec()}}.
}
