% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-trim.R
\name{str_trim_anything}
\alias{str_trim_anything}
\alias{trim_anything}
\title{Trim something other than whitespace.}
\usage{
str_trim_anything(...)

trim_anything(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_trim_anything]{strex::str_trim_anything()}}.
}
