% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{StrElem}
\alias{StrElem}
\title{Extract a single character of a string, using its index.}
\usage{
StrElem(string, index)
}
\arguments{
\item{string}{A string.}

\item{index}{An integer. Negative indexing is allowed as in
\code{\link[stringr]{str_sub}}.}
}
\value{
A one-character string.
}
\description{
Extract a single character of a string, using its index.
}
\examples{
StrElem("abcd", 3)
StrElem("abcd", -2)
}
