% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{StringToVec}
\alias{StringToVec}
\title{Convert a string to a vector of characters}
\usage{
StringToVec(string)
}
\arguments{
\item{string}{A string.}
}
\value{
A character vector.
}
\description{
Go from a string to a vector whose \eqn{i}th element is the \eqn{i}th character in the string.
}
\examples{
StringToVec("abcdef")
}

