% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StrListRemoveEmpties}
\alias{StrListRemoveEmpties}
\title{Remove empty strings from a character list.}
\usage{
StrListRemoveEmpties(char_list)
}
\arguments{
\item{char_list}{A list of character vectors.}
}
\value{
A list of character vectors.
}
\description{
Remove empty strings from a character list.
}
\examples{
StrListRemoveEmpties(list(c("a", "", "b"), "gg", c("", 1, "")))
}

