% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{StrReverse}
\alias{StrReverse}
\title{Reverse a string.}
\usage{
StrReverse(string)
}
\arguments{
\item{string}{A string.}
}
\value{
A string.
}
\description{
Reverse the order of a string; put it backwards.
}
\examples{
StrReverse("abcdef")
}

