% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{StrSplitByNums}
\alias{StrSplitByNums}
\title{Split a string by its numeric charachters.}
\usage{
StrSplitByNums(string, decimals = FALSE, leading.decimals = FALSE,
  negs = FALSE)
}
\arguments{
\item{string}{A string.}
}
\description{
Break a string wherever you go from a numeric charachter to a non-numeric or
vice-versa.
}
\examples{
StrSplitByNums("abc123def456.789gh")
StrSplitByNums("abc123def456.789gh", decimals = TRUE)
}

