% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-trim.R
\name{trim_anything}
\alias{trim_anything}
\title{Trim something other than whitespace}
\usage{
trim_anything(string, pattern, side = "both")
}
\arguments{
\item{string}{A string.}

\item{pattern}{A string. The pattern to be trimmed (\emph{not} interpreted as
regular expression). So to trim a period, use \code{char = "."} and not
\code{char = "\\\."}).}

\item{side}{Which side do you want to trim from? \code{"both"} is the
default, but you can also have just either \code{"left"} or \code{"right"}
(or optionally the shortened \code{"b"}, \code{"l"} and \code{"r"}).}
}
\description{
See \code{\link[strex:str_trim_anything]{strex::str_trim_anything()}}.
}
