% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{BeforeLastDot}
\alias{BeforeLastDot}
\title{Get the part of a string before the last period.}
\usage{
BeforeLastDot(strings)
}
\arguments{
\item{strings}{A character vector.}
}
\value{
A string.
}
\description{
This is usually used to get the part of a file name that doesn't include the file extension.
}
\examples{
BeforeLastDot(c("spreadsheet1.csv", "doc2.doc"))

}
