% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can-be-num.R
\name{can_be_numeric}
\alias{can_be_numeric}
\alias{str_can_be_numeric}
\title{Check if a string could be considered as numeric.}
\usage{
can_be_numeric(...)

str_can_be_numeric(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_can_be_numeric]{strex::str_can_be_numeric()}}.
}
