% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-currency.R
\name{currency}
\alias{currency}
\alias{str_extract_currencies}
\alias{extract_currencies}
\alias{str_nth_currency}
\alias{nth_currency}
\alias{str_first_currency}
\alias{first_currency}
\alias{str_last_currency}
\alias{last_currency}
\title{Get the currencies of numbers within a string.}
\usage{
str_extract_currencies(string)

extract_currencies(string)

str_nth_currency(string, n)

nth_currency(string, n)

str_first_currency(string)

first_currency(string)

str_last_currency(string)

last_currency(string)
}
\arguments{
\item{string}{A character vector.}

\item{n}{A vector of integerish values. Must be either length 1 or
have length equal to the length of \code{string}. Negative indices count from
the back: while \code{n = 1} and \code{n = 2} correspond to first and second, \code{n = -1} and \code{n = -2} correspond to last and second-last. \code{n = 0} will return
\code{NA}.}
}
\description{
See \code{\link[strex:str_extract_currencies]{strex::str_extract_currencies()}}.
}
