% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-currency.R
\name{currency}
\alias{currency}
\alias{get_currencies}
\alias{get_currency}
\title{Get the currencies of numbers within a string.}
\usage{
get_currencies(string)

get_currency(string)
}
\arguments{
\item{string}{A string for \code{get_currencies()} and a character vector for
\code{get_currency()}.}
}
\description{
See \code{\link[strex:str_get_currency]{strex::str_get_currency()}}.
}
