% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{ExtractNumbers}
\alias{ExtractNonNumerics}
\alias{ExtractNumbers}
\title{Extract numbers (or non-numbers) from a string.}
\usage{
ExtractNumbers(string, leave.as.string = FALSE, decimals = FALSE,
  leading.decimals = FALSE, negs = FALSE, err.na = FALSE)

ExtractNonNumerics(string, decimals = FALSE, leading.decimals = FALSE,
  negs = FALSE, err.na = FALSE)
}
\arguments{
\item{string}{A string.}

\item{leave.as.string}{Do you want to return the number as a string
(\code{TRUE}) or as numeric (\code{FALSE}, the default)?}

\item{decimals}{Do you want to include the possibility of decimal numbers
(\code{TRUE}) or not (\code{FALSE}, the default).}

\item{leading.decimals}{Do you want to allow a leading decimal point to be
the start of a number?}

\item{negs}{Do you want to allow negative numbers? Note that double negatives
are not handled here (see the examples).}

\item{err.na}{If the function can't find what you're looking for, if
\code{err.na = FALSE} (the default), the function wil stop with an error,
whereas if \code{err.na = TRUE}, the function will return \code{NA}.}
}
\description{
Extract the numbers (or non-numbers) from a string where decimals are
optionally allowed. Please view the examples at the bottom of this page to
ensure that you understand how this function works, and its limitations.
}
\examples{
ExtractNumbers("abc123abc456")
ExtractNumbers("abc1.23abc456")
ExtractNumbers("abc1.23abc456", decimals = TRUE)
ExtractNumbers("abc1..23abc456", decimals = TRUE)
ExtractNumbers("abc1..23abc456", decimals = TRUE, leading.decimals = TRUE)
ExtractNumbers("-123abc456")
ExtractNumbers("-123abc456", negs = TRUE)
ExtractNumbers("--123abc456", negs = TRUE)
ExtractNumbers("abc1.2.3", decimals = TRUE, err.na = TRUE)
ExtractNonNumerics("abc123abc456")
ExtractNonNumerics("abc1.23abc456")
ExtractNonNumerics("abc1.23abc456", decimals = TRUE)
ExtractNonNumerics("abc1..23abc456", decimals = TRUE)
ExtractNonNumerics("abc1..23abc456", decimals = TRUE, leading.decimals = TRUE)
ExtractNonNumerics("-123abc456")
ExtractNonNumerics("-123abc456", negs = TRUE)
\dontrun{
ExtractNumbers("abc1.2.3", decimals = TRUE)}
}

