% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports_deprecated.R, R/files_deprecated.R,
%   R/filesstrings.R, R/strings_deprecated.R, R/utils_deprecated.R
\name{filesstrings-defunct}
\alias{filesstrings-defunct}
\alias{PasteCollapseListElems}
\alias{StrListRemoveEmpties}
\alias{CharListElemsNthElem}
\alias{NumListElemsNthElem}
\alias{CreateDirsIfNotThere}
\alias{RemoveDirs}
\alias{MergeTablesOnDisk}
\alias{MoveFiles}
\alias{NiceFileNums}
\alias{RemoveFileNameSpaces}
\alias{RenameWithNums}
\alias{UnitDirs}
\alias{filesstrings-defunct}
\alias{CanBeNumeric}
\alias{GetCurrencies}
\alias{GetCurrency}
\alias{DuplicatesToSingles}
\alias{NiceNums}
\alias{ExtractNumbers}
\alias{ExtractNonNumerics}
\alias{NthNumber}
\alias{NthNonNumeric}
\alias{StrSplitByNums}
\alias{StrElem}
\alias{StrElemsPasted}
\alias{StringToVec}
\alias{StringsWithPatterns}
\alias{StrAfterNth}
\alias{StrBeforeNth}
\alias{BeforeLastDot}
\alias{ExtendCharVec}
\alias{PasteDifferentLengths}
\alias{PutInPos}
\alias{TrimAnything}
\alias{CountMatches}
\alias{LocateBraces}
\alias{RemoveQuoted}
\alias{GiveExt}
\alias{SplitCamelCase}
\alias{AllEqual}
\alias{GroupClose}
\title{Defunct functions}
\usage{
PasteCollapseListElems(...)

StrListRemoveEmpties(...)

CharListElemsNthElem(...)

NumListElemsNthElem(...)

CreateDirsIfNotThere(...)

RemoveDirs(...)

MergeTablesOnDisk(...)

MoveFiles(...)

NiceFileNums(...)

RemoveFileNameSpaces(...)

RenameWithNums(...)

UnitDirs(...)

CanBeNumeric(...)

GetCurrencies(...)

GetCurrency(...)

DuplicatesToSingles(...)

NiceNums(...)

ExtractNumbers(...)

ExtractNonNumerics(...)

NthNumber(...)

NthNonNumeric(...)

StrSplitByNums(...)

StrElem(...)

StrElemsPasted(...)

StringToVec(...)

StringsWithPatterns(...)

StrAfterNth(...)

StrBeforeNth(...)

BeforeLastDot(...)

ExtendCharVec(...)

PasteDifferentLengths(...)

PutInPos(...)

TrimAnything(...)

CountMatches(...)

LocateBraces(...)

RemoveQuoted(...)

GiveExt(...)

SplitCamelCase(...)

AllEqual(...)

GroupClose(...)
}
\arguments{
\item{...}{Defunct function arguments.}
}
\description{
These functions have been made defunct, mostly because the naming style of
the package has been changed. Some have been removed because they were not
well-done.
}
