% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesstrings.R
\name{filesstrings-defunct}
\alias{filesstrings-defunct}
\title{Defunct functions}
\description{
These functions have been made defunct, mostly because the naming style of
the package has been changed. To figure out which function to run instead,
run the defunct function of interest and you'll get an error message
telling you which function to use instead.
}
