% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesstrings-defunct.R
\name{filesstrings-defunct}
\alias{filesstrings-defunct}
\alias{str_with_patterns}
\alias{count_matches}
\title{Defunct functions in \code{filesstrings}}
\usage{
str_with_patterns(...)

count_matches(...)
}
\arguments{
\item{...}{Arguments to defunct functions.}
}
\description{
These functions are gone, no longer available.
}
