% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesstrings-package.R
\docType{package}
\name{filesstrings}
\alias{filesstrings}
\alias{filesstrings-package}
\title{\code{filesstrings}: handy file and string manipulation}
\description{
This started out as a package for file and string manipulation. Since then,
the \code{fs} file manipulation package and the `strex string manipulation package
emerged, offering functionality previously given by this package (but
slightly better). Those packages have hence almost pushed 'filesstrings' into
extinction. However, it still has a small number of unique, handy file
manipulation functions which can be seen in the
\href{https://cran.r-project.org/package=filesstrings/vignettes/files.html}{vignette}.
One example is a function to remove spaces from all file names in a
directory.
}
\references{
Rory Nolan and Sergi Padilla-Parra (2017). filesstrings: An R
package for file and string manipulation. The Journal of Open Source
Software, 2(14).  \doi{10.21105/joss.00260}.
}
