% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-give-ext.R
\name{give_ext}
\alias{give_ext}
\title{Ensure a file name has the intended extension.}
\usage{
give_ext(string, ext, replace = FALSE)
}
\arguments{
\item{string}{The intended file name.}

\item{ext}{The intended file extension (with or without the ".").}

\item{replace}{If the file has an extension already, replace it (or append
the new extension name)?}
}
\description{
See \code{\link[strex:str_give_ext]{strex::str_give_ext()}}.
}
