% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-locate.R
\name{locate_braces}
\alias{locate_braces}
\alias{str_locate_braces}
\title{Locate the braces in a string.}
\usage{
locate_braces(string)

str_locate_braces(string)
}
\arguments{
\item{string}{A character vector}
}
\description{
See \code{\link[strex:str_locate_braces]{strex::str_locate_braces()}}.
}
