% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{MergeTables}
\alias{MergeTables}
\title{Merge Tables.}
\usage{
MergeTables(file.names, delim, out.name, header = TRUE)
}
\arguments{
\item{file.names}{The paths to the files to merge.}

\item{delim}{Delimeter used to separate values.}

\item{out.name}{The path to the output file containing the merged tables.}

\item{header}{Do the tables to be merged have headers?}
}
\description{
Merge tables saved on disk as delimited files.
}
