% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-nice-nums.R
\name{nice_nums}
\alias{nice_nums}
\title{Make string numbers comply with alphabetical order}
\usage{
nice_nums(strings)
}
\arguments{
\item{strings}{A vector of strings.}
}
\description{
See \code{\link[strex:str_alphord_nums]{strex::str_alphord_nums()}}.
}
