% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-num-after.R
\name{nth_number_after_mth}
\alias{nth_number_after_mth}
\alias{str_nth_number_after_mth}
\alias{nth_number_after_first}
\alias{nth_number_after_last}
\alias{first_number_after_mth}
\alias{last_number_after_mth}
\alias{first_number_after_first}
\alias{first_number_after_last}
\alias{last_number_after_first}
\alias{last_number_after_last}
\alias{str_nth_number_after_first}
\alias{str_nth_number_after_last}
\alias{str_first_number_after_mth}
\alias{str_last_number_after_mth}
\alias{str_first_number_after_first}
\alias{str_first_number_after_last}
\alias{str_last_number_after_first}
\alias{str_last_number_after_last}
\title{Find the \code{n}th number after the \code{m}th occurrence of a pattern.}
\usage{
nth_number_after_mth(...)

str_nth_number_after_mth(...)

nth_number_after_first(...)

nth_number_after_last(...)

first_number_after_mth(...)

last_number_after_mth(...)

first_number_after_first(...)

first_number_after_last(...)

last_number_after_first(...)

last_number_after_last(...)

str_nth_number_after_first(...)

str_nth_number_after_last(...)

str_first_number_after_mth(...)

str_last_number_after_mth(...)

str_first_number_after_first(...)

str_first_number_after_last(...)

str_last_number_after_first(...)

str_last_number_after_last(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_nth_number_after_mth]{strex::str_nth_number_after_mth()}}.
}
