% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{RemoveDirs}
\alias{RemoveDirs}
\title{Remove directories.}
\usage{
RemoveDirs(dirs)
}
\arguments{
\item{dirs}{The names of the directories, specified via relative or absolute
paths.}
}
\value{
Invisibly, a logical vector with \code{TRUE} for each success and
\code{FALSE} for failures. See \link[base:unlink]{base::unlink}.
}
\description{
Delete directories and all of their contents (can just be one directory).
}
\examples{
sapply(c("mydir1", "mydir2"), dir.create)
RemoveDirs(c("mydir1", "mydir2"))
}
