% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-singleize.R
\name{singleize}
\alias{singleize}
\title{Remove back-to-back duplicates of a pattern in a string.}
\usage{
singleize(string, pattern)
}
\arguments{
\item{string}{A character vector. The string(s) to be purged of duplicates.}

\item{pattern}{A character vector. Pattern(s) specified like the pattern(s)
in the stringr package (e.g. look at \code{\link[stringr:str_locate]{stringr::str_locate()}}). If
this has length >1 its length must be the same as that of \code{string}.}
}
\description{
See \code{\link[strex:str_singleize]{strex::str_singleize()}}.
}
