% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{SplitCamelcase}
\alias{SplitCamelcase}
\title{Split a string based on camelcase}
\usage{
SplitCamelcase(string, lower = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{lower}{Do you want the output to be all lower case (or as is)?}
}
\value{
A list of character vectors, one list element for each element of \code{string}.
}
\description{
Vectorised over \code{string}.
}
\examples{
SplitCamelcase(c("RoryNolan", "NaomiFlagg", "DepartmentOfSillyHats"))
}
\references{
Adapted from Ramnath Vaidyanathan's answer at http://stackoverflow.com/questions/8406974/splitting-camelcase-in-r.
}

