% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-after.R, R/str-before.R
\name{str_after_nth}
\alias{str_after_nth}
\alias{str_after_first}
\alias{str_after_last}
\alias{str_before_nth}
\alias{str_before_first}
\alias{str_before_last}
\title{Text before or after \eqn{n}th occurrence of pattern.}
\usage{
str_after_nth(strings, pattern, n)

str_after_first(strings, pattern)

str_after_last(strings, pattern)

str_before_nth(strings, pattern, n)

str_before_first(strings, pattern)

str_before_last(strings, pattern)
}
\arguments{
\item{strings}{A character vector.}

\item{pattern}{A character vector. Pattern(s) specified like the pattern(s)
in the stringr package (e.g. look at \code{\link[stringr:str_locate]{stringr::str_locate()}}). If
this has length >1 its length must be the same as that of \code{string}.}

\item{n}{A natural number to identify the \eqn{n}th occurrence (defaults to
first (\code{n = 1})). This can be negatively indexed, so if you wish to select
the \emph{last} occurrence, you need \code{n = -1}, for the second-last, you
need \code{n = -2} and so on.}
}
\description{
See \code{\link[strex:str_after_nth]{strex::str_after_nth()}}.
}
