% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-after.R
\name{str_after_nth}
\alias{str_after_nth}
\alias{after_nth}
\alias{str_after_first}
\alias{after_first}
\alias{str_after_last}
\alias{after_last}
\title{Text after the \code{n}th occurrence of pattern.}
\usage{
str_after_nth(...)

after_nth(...)

str_after_first(...)

after_first(...)

str_after_last(...)

after_last(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copies of \code{\link[strex:before-and-after]{strex::str_after_nth()}} and friends.
}
