% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-after.R
\name{str_after_nth}
\alias{str_after_nth}
\alias{after_nth}
\alias{str_after_first}
\alias{after_first}
\alias{str_after_last}
\alias{after_last}
\title{Text after the \code{n}th occurrence of pattern.}
\usage{
str_after_nth(string, pattern, n)

after_nth(string, pattern, n)

str_after_first(string, pattern)

after_first(string, pattern)

str_after_last(string, pattern)

after_last(string, pattern)
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{The pattern to look for.

The default interpretation is a regular expression, as described in
\link[stringi:about_search_regex]{stringi::about_search_regex}.

To match a without regular expression (i.e. as a human would), use
\link[stringr:modifiers]{coll()}. For details see \code{\link[stringr:modifiers]{stringr::regex()}}.}

\item{n}{A vector of integerish values. Must be either length 1 or
have length equal to the length of \code{string}. Negative indices count from
the back: while \code{n = 1} and \code{n = 2} correspond to first and second, \code{n = -1} and \code{n = -2} correspond to last and second-last. \code{n = 0} will return
\code{NA}.}
}
\description{
See \code{\link[strex:before-and-after]{strex::str_after_nth()}}.
}
