% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-before.R
\name{str_before_nth}
\alias{str_before_nth}
\alias{before_nth}
\alias{str_before_first}
\alias{before_first}
\alias{str_before_last}
\alias{before_last}
\title{Text before the \code{n}th occurrence of pattern.}
\usage{
str_before_nth(...)

before_nth(...)

str_before_first(...)

before_first(...)

str_before_last(...)

before_last(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copies of \code{\link[strex:before-and-after]{strex::str_before_nth()}} and friends.
}
