% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-elem.R
\name{str_elem}
\alias{str_elem}
\alias{elem}
\title{Extract a single character from a string, using its index.}
\usage{
str_elem(...)

elem(...)
}
\arguments{
\item{...}{Pass-through to \code{strex} function.}
}
\description{
Copy of \code{\link[strex:str_elem]{strex::str_elem()}}.
}
