% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-elem.R
\name{str_elem}
\alias{str_elem}
\title{Extract a single character from a string, using its index.}
\usage{
str_elem(string, index)
}
\arguments{
\item{string}{A string.}

\item{index}{An integer. Negative indexing is allowed as in
\code{\link[stringr:str_sub]{stringr::str_sub()}}.}
}
\description{
See \code{\link[strex:str_elem]{strex::str_elem()}}.
}
