% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-locate.R
\name{str_nth_instance_indices}
\alias{str_nth_instance_indices}
\alias{str_first_instance_indices}
\alias{str_last_instance_indices}
\title{Get the indices of the \eqn{n}th instance of a pattern.}
\usage{
str_nth_instance_indices(string, pattern, n)

str_first_instance_indices(string, pattern)

str_last_instance_indices(string, pattern)
}
\arguments{
\item{string}{A character vector. These functions are vectorized over this
argument.}

\item{pattern}{A character vector. Pattern(s) specified like the pattern(s)
in the stringr package (e.g. look at \code{\link[stringr:str_locate]{stringr::str_locate()}}). If
this has length >1 its length must be the same as that of \code{string}.}

\item{n}{Then \eqn{n} for the \eqn{n}th instance of the pattern.}
}
\description{
See \code{\link[strex:str_locate_nth]{strex::str_locate_nth()}}.
}
