% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-elem.R
\name{str_paste_elems}
\alias{str_paste_elems}
\alias{paste_elems}
\title{Extract bits of a string and paste them together.}
\usage{
str_paste_elems(string, indices, sep = "")

paste_elems(string, indices, sep = "")
}
\arguments{
\item{string}{A character vector.}

\item{indices}{A vector of integerish values. Negative indexing is allowed as
in \code{\link[stringr:str_sub]{stringr::str_sub()}}.}

\item{sep}{A string. The separator for pasting \code{string} elements together.}
}
\description{
See \code{\link[strex:str_paste_elems]{strex::str_paste_elems()}}.
}
