% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-split-by-nums.R
\name{str_split_by_nums}
\alias{str_split_by_nums}
\title{Split a string by its numeric characters.}
\usage{
str_split_by_nums(string, decimals = FALSE, leading_decimals = FALSE,
  negs = FALSE)
}
\arguments{
\item{string}{A string.}

\item{decimals}{Do you want to include the possibility of decimal numbers
(\code{TRUE}) or not (\code{FALSE}, the default).}

\item{leading_decimals}{Do you want to allow a leading decimal point to be
the start of a number?}

\item{negs}{Do you want to allow negative numbers? Note that double negatives
are not handled here (see the examples).}
}
\description{
See \code{\link[strex:str_split_by_nums]{strex::str_split_by_nums()}}.
}
