% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-to-vec.R
\name{str_to_vec}
\alias{str_to_vec}
\title{Convert a string to a vector of characters}
\usage{
str_to_vec(string)
}
\arguments{
\item{string}{A string.}
}
\description{
See \code{\link[strex:str_to_vec]{strex::str_to_vec()}}.
}
