% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{StrSplitByNums}
\alias{StrSplitByNums}
\title{Split a string by its numeric charachters.}
\usage{
StrSplitByNums(string, decimals = FALSE, leading.decimals = FALSE,
  negs = FALSE)
}
\arguments{
\item{string}{A string.}

\item{decimals}{Do you want to include the possibility of decimal numbers
(\code{TRUE}) or not (\code{FALSE}, the default).}

\item{leading.decimals}{Do you want to allow a leading decimal point to be
the start of a number?}

\item{negs}{Do you want to allow negative numbers? Note that double negatives
are not handled here (see the examples).}
}
\description{
Break a string wherever you go from a numeric charachter to a non-numeric or
vice-versa.
}
\examples{
StrSplitByNums(c("abc123def456.789gh", "a1b2c344"))
StrSplitByNums("abc123def456.789gh", decimals = TRUE)
}

