% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-trim.R
\name{trim_anything}
\alias{trim_anything}
\alias{str_trim_anything}
\title{Trim something other than whitespace}
\usage{
trim_anything(string, pattern, side = "both")

str_trim_anything(string, pattern, side = "both")
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{The pattern to look for.

The default interpretation is a regular expression, as described in
\link[stringi:about_search_regex]{stringi::stringi-search-regex}.

To match a without regular expression (i.e. as a human would), use
\link[stringr:modifiers]{coll()}. For details see \code{\link[stringr:modifiers]{stringr::regex()}}.}

\item{side}{Which side do you want to trim from? \code{"both"} is the
default, but you can also have just either \code{"left"} or \code{"right"}
(or optionally the shortened \code{"b"}, \code{"l"} and \code{"r"}).}
}
\description{
See \code{\link[strex:str_trim_anything]{strex::str_trim_anything()}}.
}
